Make clear the responsibilities of the Reactant, its Receiver and Inbox regarding the stimulation of the Reactant.
The Receiver should stimulate the Reactant only when the Inbox becomes nonEmpty.
The Reactant should have a single flag (isReadyToProcessMsg) that determines if stimulus are ignored and this flag transition to the state that accepts stimulus should be done after ensuring all pending messages were processed in order to compensate the ignored stimulus.