Finish the implementation of `TimedAssistantProvider` and add `TimedDoerProvider` and `MatrixTimedDoer` to support scheduling.
Also rename to third person tense the methods that return a `Duty`/`Task` and require the returned `Duty`/`Task` be triggered in order for the method's intended action to be performed.

Add the `PruebaTimed` to test the scheduling work. Curiously it had a single bug (a `wait` that had to be an `await`). I can't believe it!
Surely there are more bugs I will find later, as always.