Add more requirements to the client and state machine of `ConsensusParticipant`.
Client commands must carry enough information to support deduplication, ordering, and conflict detection. Typically, this includes a `ClientId` and a monotonically increasing request identifier or timestamp.
If a client replies is inconsistent, the service would respond with `InconsistentState`,
The responsibility of supporting the above is delegated to the `ConsensusParticipantSdm.Workspace` implementation by means of the `indexOfLastAppendedCommandFrom(clientId: ClientId)` method.
Add the notification of commit-index changes to be able to test the Leader completeness invariant.

Correct the bugs:
- make `onCommandFromClient` be idempotent as told above.
- in `onAppendRecords` when there is the `prevRecordIndex` is greater than the index of the last appended record.

- lastAppliedCommandIndex was not updated in some places.

- stop calling setMessagesListener(null) and setting workspace to null in when the service is stopped to avoid forcing the implementation of the required interfaces to do null checks.
- the leader election criteria to prioritize term over be leader or not over log state, because that information may be stale.